% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3_boundary.R
\name{m3_b2_grid}
\alias{m3_b2_grid}
\title{Estimate b2 based on asymptotic distribution}
\usage{
m3_b2_grid(tps = seq(0.5, 0.8, by = 0.05), n_large = 1e+06, ...)
}
\description{
Estimate b2 based on asymptotic distribution
}
