% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3_boundary.R
\name{m3_calculate_b2}
\alias{m3_calculate_b2}
\title{Get boundary of b2}
\usage{
m3_calculate_b2(
  tp,
  est_b2 = rbind(c(0.65, 2.58), c(0.7, 2.43), c(0.75, 2.37)),
  type = "asP",
  ...
)
}
\description{
Get boundary of b2
}
