% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3_simu.R
\name{m3_get_cut_caltime}
\alias{m3_get_cut_caltime}
\title{Find calendar time for given number of events}
\usage{
m3_get_cut_caltime(dta_surv, n_event = Inf, arms = 1:3)
}
\description{
Find calendar time for given number of events
}
