% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3_testing.R
\name{m3_rej_trial_single}
\alias{m3_rej_trial_single}
\title{Test for an individual trial}
\usage{
m3_rej_trial_single(
  zs,
  tp,
  ob_type = "asOF",
  b2_type = c("asP", "optimal"),
  alpha = 0.0125,
  ...
)
}
\description{
Test for an individual trial
}
