% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3_simu.R
\name{m3_simu_pts}
\alias{m3_simu_pts}
\title{Simulate patients}
\usage{
m3_simu_pts(
  median_mth_pla = 5,
  arms_hr = c(1, 1),
  arms_randr = c(1, 1),
  drop_annual = 0.05,
  enroll_duration = 342/21,
  ntot = 342,
  ntot_event = 157
)
}
\description{
Simulate patients
}
