% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3_simu.R
\name{m3_simu_trial_single}
\alias{m3_simu_trial_single}
\title{Simulate a trial and get z-sccores}
\usage{
m3_simu_trial_single(
  ...,
  ntot_event = 157,
  int_nevent_arm_pt = 111,
  int_nevent_arm_p = 70,
  tp = NULL
)
}
\description{
Get z-scores for one interim analysis
}
